+++ src/app/providers/index.ts
@@ -1,4 +1,4 @@
 export * from './feedback.provider';
-export * from './list-files.providers';
-export * from './list-modules.providers';
-export * from './list-routes.providers';
+export * from './list-files.provider';
+export * from './list-modules.provider';
+export * from './list-routes.provider';